// src/features/Alerts/Mounting/types/mountingTypes.ts
export type { MountingResponse };

interface MountingResponse {
  number_alert: number;
  subsidiary_id: number;
  company_id: number;
  corporate_id: number;
  ranking_alert: string;
  alert_type: string;
  priority: string;
  status: number;
  economic_number: string;
  vehicle_brand_id: number;
  vehicle_type_id: number;
  alert_vehicle_tire_id: number;
  vehicle_tire_id: number;
  vehicle_id: number;
  vehicle_brand: string;
  vehicle_type: string;
  driver_id: number;
  search: string;
  created_at: string;
  updated_at: string;
  driver_name: string;
  division_name: string;
  subsidiary_name: string;
  company_name: string;
  corporate_name: string;
}